+++ src/lib/task-generators/index.ts
@@ -1,6 +1,0 @@
-import * as clean from "./clean";
-import * as copy from "./copy";
-import * as pug from "./pug";
-import * as sass from "./sass";
-
-export { clean, copy, pug, sass };